Openapi REST client fixes

+++ src/community/rest-openapi/java-client/src/main/java/org/geoserver/restconfig/client/LayersClient.java
@@ -22,6 +22,7 @@ public class LayersClient {
         return client.api(LayersApi.class);
     }
 
+    /** Adapter for {@link LayersApi#getLayer(String)} */
     public Optional<Layer> getLayer(@NonNull String qualifiedLayerName) {
         try {
             return Optional.of(api().getLayer(qualifiedLayerName).getLayer());
@@ -30,37 +31,81 @@ public class LayersClient {
         }
     }
 
+    /** Adapter for {@link LayersApi#getLayerByWorkspace} */
     public Optional<Layer> getLayer(@NonNull String workspace, @NonNull String layerName) {
         try {
+            // tell the server not to flood the logs if not found, as odd as it seems
+            Boolean quietOnNotFound = Boolean.TRUE;
             return Optional.of(
-                    api().getLayerByWorkspace(workspace, layerName, Boolean.TRUE).getLayer());
+                    api().getLayerByWorkspace(workspace, layerName, quietOnNotFound).getLayer());
         } catch (ServerException.NotFound nf) {
             return Optional.empty();
         }
     }
 
+    /** Adapter for {@link LayersApi#getLayers} */
     public List<NamedLink> getLayers() {
         Layers layers = api().getLayers();
         LayerReference ref = layers.getLayers();
         return ref.getLayer();
     }
 
+    /** Adapter for {@link LayersApi#getLayersByWorkspace} */
     public List<NamedLink> getLayers(@NonNull String workspaceName) {
         Layers layers = api().getLayersByWorkspace(workspaceName);
         LayerReference ref = layers.getLayers();
         return ref.getLayer();
     }
 
+    /** Adapter for {@link LayersApi#deleteLayer(String, Boolean)} with {@code recurse = false} */
     public void deleteLayer(@NonNull String qualifiedLayerName) {
+        if (!qualifiedLayerName.contains(":")) {
+            throw new IllegalArgumentException(
+                    "Layer name must be qualified (prefixed with workspace name, e.g. ws:name");
+        }
         boolean recurse = true;
         api().deleteLayer(qualifiedLayerName, recurse);
     }
 
+    /**
+     * Recursively removes the layer from all layer groups which reference it. If this results in an
+     * empty layer group, also deletes the layer group.
+     *
+     * <p>Adapter for {@link LayersApi#deleteLayer(String, Boolean)} with {@code recurse = true}
+     */
+    public void deleteLayerRecursively(@NonNull String qualifiedLayerName) {
+        if (!qualifiedLayerName.contains(":")) {
+            throw new IllegalArgumentException(
+                    "Layer name must be qualified (prefixed with workspace name, e.g. ws:name");
+        }
+        boolean recurse = true;
+        api().deleteLayer(qualifiedLayerName, recurse);
+    }
+
+    /**
+     * Adapter for {@link LayersApi#deleteLayerByWorkspace(String, String, Boolean)} with {@code
+     * recurse = false}
+     */
     public void deleteLayer(@NonNull String workspaceName, @NonNull String layerName) {
+        boolean recurse = false;
+        api().deleteLayerByWorkspace(workspaceName, layerName, recurse);
+    }
+
+    /**
+     * Recursively removes the layer from all layer groups which reference it. If this results in an
+     * empty layer group, also deletes the layer group.
+     *
+     * <p>Adapter for {@link LayersApi#deleteLayerByWorkspace(String, String, Boolean)} with {@code
+     * recurse = true}
+     */
+    public void deleteLayerRecursively(@NonNull String workspaceName, @NonNull String layerName) {
         boolean recurse = true;
         api().deleteLayerByWorkspace(workspaceName, layerName, recurse);
     }
 
+    /**
+     * Adapter for {@link LayersApi#updateLayerByWorkspace(String, String, LayerInfoWrapper))
+     */
     public void updateLayer(
             @NonNull String workspaceName,
             @NonNull String layerName,
